After managing product and sending form user is redirected to productsList

+++ src/main/java/com/devshop/devshop/repository/ProductRepository.java
@@ -2,9 +2,10 @@ package com.devshop.devshop.repository;
 
 import com.devshop.devshop.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
     List<Product> findByCategoryId(Long id);
 }
